fixing issue wiht reading from buffer

+++ src/main/java/org/apache/druid/extensions/aggregators/multivalue/MultiValueContainsAggregatorFactory.java
@@ -173,7 +173,7 @@ public class MultiValueContainsAggregatorFactory extends AggregatorFactory
   @Override
   public int getMaxIntermediateSize()
   {
-    return Integer.BYTES;
+    return Long.BYTES;
   }
 
   @Override
